refactor: split BuildingTools into BuildingQueryTools + BuildingDeconstructTools (#169)
